able to get info

+++ src/types/info.rs
@@ -1,3 +1,5 @@
+use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
+
 
 #[derive(Deserialize, Serialize, Debug, PartialEq)]
 pub struct Info {
@@ -19,12 +21,12 @@ pub struct Info {
     pub tiptime: u32,
     pub connections: u32,
     pub proxy: String,
-    pub difficulty: u32,
+    pub difficulty: f64,
     pub testnet: bool,
     pub keypoololdest: u32,
     pub keypoolsize: u32,
-    pub paytxfee: u32,
-    pub relayfee: u32,
-    pub errors: String, // could be standardized
+    pub paytxfee: f32,
+    pub relayfee: f32,
+    pub errors: String,
     pub name: String,
 }
